+++ src/index.js
@@ -53,12 +53,82 @@ function checkRepo(repo) {
   assert(/^[A-Za-z0-9_.-]+\/[A-Za-z0-9_.-]+$/.test(repo), `Invalid repository name: ${repo}`);
 }
 
+function defaultOnError(sha, email, error) {
+  // eslint-disable-next-line no-console
+  console.error(`Failed to get the author of ${sha} with the email <${email}>: ${error}`);
+}
+
+function getErrorCallback(onerror) {
+  if (onerror === 'undefined') return defaultOnError;
+  if (typeof onerror === 'function') return onerror;
+  return () => {};
+}
+
+function createSetApiPromiseQueue(configs) {
+  const { repo, token = process.env.GITHUB_TOKEN, onerror } = configs;
+  checkRepo(repo);
+  const errorCallback = getErrorCallback(onerror);
+  const promiseForEmail = new Map();
+  return {
+    setApiPromiseQueue: async.queue(async ({
+      email, sha,
+    }) => {
+      if (!promiseForEmail.has(email)) {
+        promiseForEmail.set(email, axios.get(`https://api.github.com/repos/${repo}/commits/${sha}`, {
+          headers: {
+            Accept: 'application/vnd.github.v3+json',
+            ...(typeof token === 'string' && { Authorization: `token ${token}` }),
+          },
+        }).catch((error) => errorCallback(sha, email, error)));
+      }
+    }),
+    promiseForEmail,
+  };
+}
+
 /**
  * @callback OnErrorCallback
  * @param {string} [sha] - The sha of the commit which caused the error.
  * @param {string} [email] - The email of the author to the commit which caused the error.
  * @param {any} [error] - The error.
  */
+
+async function getAuthorsWithSetApiPromiseQueue(configs, setApiPromiseQueue, promiseForEmail) {
+  const {
+    filePath,
+    cache = new Map(),
+    follow = true,
+    concurrency = 64,
+    onerror,
+    git = 'git',
+  } = configs;
+  const errorCallback = getErrorCallback(onerror);
+  const absPath = path.resolve(filePath);
+  const fileStat = await fs.lstat(absPath);
+  const { stdout } = await exec(`${git} log ${follow ? '--follow' : '--no-follow'} --pretty='%H %aE' ${absPath}`, {
+    cwd: fileStat.isDirectory() ? absPath : path.dirname(absPath),
+  });
+  const authors = new Set();
+  const queue = async.queue(async (line) => {
+    const parts = line.split(' ');
+    if (parts.length !== 2) return;
+    const [sha, email] = parts;
+    if (cache.has(email)) {
+      authors.add(cache.get(email));
+      return;
+    }
+    await setApiPromiseQueue.push({ email, sha });
+    const response = await promiseForEmail.get(email);
+    if (!response) return;
+    const login = response.data?.author?.login;
+    if (typeof login === 'string') {
+      cache.set(email, login);
+      authors.add(login);
+    } else errorCallback(sha, email, '.author.login not found in the API response');
+  }, concurrency);
+  await Promise.all(stdout.split('\n').map((line) => queue.push(line)));
+  return Array.from(authors);
+}
 
 /**
  * Get the authors of a specific path in the given GitHub repository.
@@ -81,52 +151,9 @@ function checkRepo(repo) {
  * @async
  * @date 2021-02-13
  */
-async function getAuthors({
-  repo,
-  filePath,
-  token = process.env.GITHUB_TOKEN,
-  cache = new Map(),
-  follow = true,
-  concurrency = 64,
-  // eslint-disable-next-line no-console
-  onerror = (sha, email, error) => console.error(`Failed to get the author of ${sha} with the email <${email}>: ${error}`),
-  git = 'git',
-}) {
-  checkRepo(repo);
-  const errorCallback = typeof onerror === 'function' ? onerror : () => {};
-  const absPath = path.resolve(filePath);
-  const fileStat = await fs.lstat(absPath);
-  const { stdout } = await exec(`${git} log ${follow ? '--follow' : '--no-follow'} --pretty='%H %aE' ${absPath}`, {
-    cwd: fileStat.isDirectory() ? absPath : path.dirname(absPath),
-  });
-  const authors = new Set();
-  const promiseForEmail = new Map();
-  const queue = async.queue(async (line) => {
-    const parts = line.split(' ');
-    if (parts.length !== 2) return;
-    const [sha, email] = parts;
-    if (cache.has(email)) {
-      authors.add(cache.get(email));
-      return;
-    }
-    if (!promiseForEmail.has(email)) {
-      promiseForEmail.set(email, axios.get(`https://api.github.com/repos/${repo}/commits/${sha}`, {
-        headers: {
-          Accept: 'application/vnd.github.v3+json',
-          ...(typeof token === 'string' && { Authorization: `token ${token}` }),
-        },
-      }).catch((error) => errorCallback(sha, email, error)));
-    }
-    const response = await promiseForEmail.get(email);
-    if (!response) return;
-    const login = response.data?.author?.login;
-    if (typeof login === 'string') {
-      cache.set(email, login);
-      authors.add(login);
-    } else errorCallback(sha, email, '.author.login not found in the API response');
-  }, concurrency);
-  await Promise.all(stdout.split('\n').map((line) => queue.push(line)));
-  return Array.from(authors);
+async function getAuthors(configs) {
+  const { setApiPromiseQueue, promiseForEmail } = createSetApiPromiseQueue(configs);
+  return getAuthorsWithSetApiPromiseQueue(configs, setApiPromiseQueue, promiseForEmail);
 }
 
 /**
@@ -141,6 +168,7 @@ async function getAuthors({
  * not. i.e. Continue listing the history of a file beyond renames.
  * WARNING: In order to make this option work, you need to list ALL files in [paths], not only the
  * directories containing the files.
+ * @param {number=} [configs.concurrency=64] - Maximum number of tasks at the same time.
  * @param {OnErrorCallback=} [configs.onerror=console.error(...)]
  * The callback function when error happens.
  * @param {string=} [configs.git='git'] - The command (path to the binary file) for Git.
@@ -148,22 +176,16 @@ async function getAuthors({
  * @async
  * @date 2021-02-13
  */
-async function cacheFor({
-  repo,
-  paths = [],
-  token = process.env.GITHUB_TOKEN,
-  follow = true,
-  concurrency = 64,
-  // eslint-disable-next-line no-console
-  onerror = (sha, email, error) => console.error(`Failed to get the author of ${sha} with the email <${email}>: ${error}`),
-  git = 'git',
-}) {
-  checkRepo(repo);
+async function cacheFor(configs) {
+  const { paths = [] } = configs;
   assert(Array.isArray(paths));
+  const { setApiPromiseQueue, promiseForEmail } = createSetApiPromiseQueue(configs);
   const cache = new Map();
-  await Promise.all(paths.map((filePath) => getAuthors({
-    repo, filePath, token, cache, follow, concurrency, onerror, git,
-  })));
+  await Promise.all(paths.map((filePath) => getAuthorsWithSetApiPromiseQueue(
+    { ...configs, filePath, cache },
+    setApiPromiseQueue,
+    promiseForEmail,
+  )));
   return cache;
 }