+++ test/test.js
@@ -26,7 +26,7 @@ function p(...filePath) {
 describe('getAuthor', () => {
   it('test/test1 follow without cache', async () => {
     const authors = await getAuthors({ repo, filePath: p('test', 'test1'), token });
-    assert.deepEqual(authors.sort(), ['ouuan', 'sshwy']);
+    assert.deepEqual(authors, ['ouuan', 'sshwy']);
   });
   it('test/test1 no-follow without cache', async () => {
     const cache = await cacheFor({ repo, token });
@@ -34,19 +34,19 @@ describe('getAuthor', () => {
     const authors = await getAuthors({
       repo, filePath: p('test', 'test1'), follow: false, cache,
     });
-    assert.deepEqual(authors.sort(), ['sshwy']);
+    assert.deepEqual(authors, ['sshwy']);
   });
   it('test/test1 follow with fake cache', async () => {
     const authors = await getAuthors({
       repo, filePath: p('test', 'test1'), cache: new Map([fakeCache.nauuo]),
     });
-    assert.deepEqual(authors.sort(), ['nauuo', 'sshwy']);
+    assert.deepEqual(authors, ['nauuo', 'sshwy']);
   });
   it('test/test1 no-follow with fake cache', async () => {
     const authors = await getAuthors({
       repo, filePath: p('test', 'test1'), follow: false, cache: new Map([fakeCache.hwy]),
     });
-    assert.deepEqual(authors.sort(), ['hwy']);
+    assert.deepEqual(authors, ['hwy']);
   });
   it('test2/ouuan follow with updated cache', async () => {
     const cache = await cacheFor({ repo, paths: [p()] });
@@ -54,7 +54,7 @@ describe('getAuthor', () => {
     const authors = await getAuthors({
       repo, filePath: p('test2', 'ouuan'), cache,
     });
-    assert.deepEqual(authors.sort(), ['Enter-tainer', 'ouuan']);
+    assert.deepEqual(authors, ['Enter-tainer', 'ouuan']);
   });
   it('test2/ouuan no-follow with updated cache', async () => {
     const cache = await cacheFor({ repo, paths: [p('test2/ouuan'), p('test/test4')] });
@@ -62,14 +62,14 @@ describe('getAuthor', () => {
     const authors = await getAuthors({
       repo, filePath: p('test2', 'ouuan'), follow: false, cache,
     });
-    assert.deepEqual(authors.sort(), ['Enter-tainer', 'ouuan']);
+    assert.deepEqual(authors, ['Enter-tainer', 'ouuan']);
   });
   it('read cache.json', async () => {
     const cache = await readCache(path.resolve(__dirname, 'cache.json'));
     const authors = await getAuthors({
       repo, filePath: p('test/test2'), token, cache,
     });
-    assert.deepEqual(authors.sort(), ['nauuo', 'sshwy']);
+    assert.deepEqual(authors, ['nauuo', 'sshwy']);
     assert.deepEqual(cache, new Map([fakeCache.nauuo, realCache.sshwy]));
   });
   it('read/write cache', async () => {
@@ -81,6 +81,6 @@ describe('getAuthor', () => {
   });
   it('Relative path', async () => {
     const authors = await getAuthors({ repo, filePath: path.relative(process.cwd(), p('test', 'test1')), token });
-    assert.deepEqual(authors.sort(), ['ouuan', 'sshwy']);
+    assert.deepEqual(authors, ['ouuan', 'sshwy']);
   });
 });